Don't throw an exception when handling normal requests and origin is not allowed
The way it's supposed to work is that we should only throw exceptions and set HTTP status code for preflight requests, not "normal" requests. This also changes how the Access-Control-Allow-Origin header is set on responses - now it's only set to the value of Origin when the request Origin is allowed. If the Origin is not allowed, the header should be missing, which will prevent the browser from reading the request.

FIxes #53